requirements.txt
sentence_transformers
numpy
faiss-cpu
